pyproject.toml
[project]
name = "portfolio"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "strawberry-graphql",
    "psycopg[binary]",
]
